Project refactor v1, things to delete
