+++ package_party/party_manager.py
@@ -12,15 +12,3 @@ def add_player(player):
 def create_player(player_name):
     add_player(class_player.Player(len(players_list) + 1, player_name))
 
-
-def next_player_frame(player_id, current_frame):
-    players_list[player_id].next_turn(current_frame)
-
-
-def start_game():
-    global current_frame, players_list
-
-    for i in range(NUMBER_OF_FRAME):
-        for j in range(len(players_list)):
-            next_player_frame(j, i)
-
